求货

+++ application/entity/sp_member_seller.php
@@ -2,7 +2,7 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
-'uid' => 'mediumint',
+'uid' => 'int',
 'store_url' => 'varchar',
 'source_pic' => 'varchar',
 'trade_pic' => 'varchar',
