+++ application/entity/sp_pm_message.php
@@ -3,8 +3,8 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
 'id' => 'mediumint',
-'uid' => 'mediumint',
-'from_uid' => 'mediumint',
+'uid' => 'int',
+'from_uid' => 'int',
 'site_msgid' => 'mediumint',
 'msg_type' => 'tinyint',
 'readed' => 'tinyint',
